1. 修复yaml-helper.sh文件中_getAllParamPathAndValueByParentPath函数中存在的BUG。 2. 修复combineParamsToValuesYaml_ex函数存在的BUG 3. 取消grep参数中的"\}"转义，直接写成"}"

+++ script/pipeline-stages/java/docker-extend-point.sh
@@ -15,6 +15,7 @@ function _onBeforeCreatingDockerImage_ex() {
   export gDockerBuildDir
   export gProjectBuildOutDir
   export gServiceName
+  export gRuntimeVersion
 
   local l_dockerfile=$3
 
@@ -23,6 +24,10 @@ function _onBeforeCreatingDockerImage_ex() {
   local l_subDirs
   local l_subDir
   local l_content
+  local l_command
+  local l_jdkVersion
+
+  l_jdkVersion=$(grep -oE "[0-9]+" <<< "gRuntimeVersion")
 
   l_jarFiles=$(find "${gProjectBuildOutDir}" -maxdepth 1 -type f -name "${gServiceName}*.jar")
   [[ ! "${l_jarFiles[0]}" ]] &&  error "${gProjectBuildOutDir##*/}目录中不存在${gServiceName}*.jar文件"
@@ -31,15 +36,24 @@ function _onBeforeCreatingDockerImage_ex() {
   l_curDir=$(pwd)
   # shellcheck disable=SC2164
   cd "${gDockerBuildDir}"
-  if ! java -Djarmode=tools -jar "${l_jarFiles[0]}" extract --destination ./extract --layers --launcher;then
-    error "分层解压jar文件异常"
-  fi
-  info "分层解压jar文件成功"
 
-  #将./extract目录中的所有子目录复制到gDockerBuildDir目录下。
-  mv ./extract/* "${gDockerBuildDir}"
-  #删除./extract目录
-  rm -rf "./extract"
+  if [[ "${l_jdkVersion}" -ge 17 ]];then
+    l_command="java -Djarmode=tools -jar ${l_jarFiles[0]} extract --destination ./extract --layers --launcher"
+    if ! java -Djarmode=tools -jar "${l_jarFiles[0]}" extract --destination ./extract --layers --launcher;then
+      error "分层解压jar文件异常"
+    fi
+    #将./extract目录中的所有子目录复制到gDockerBuildDir目录下。
+    mv ./extract/* "${gDockerBuildDir}"
+    #删除./extract目录
+    rm -rf "./extract"
+  else
+    l_command="java -Djarmode=layertools -jar ${l_jarFiles[0]} extract"
+    if ! java -Djarmode=layertools -jar "${l_jarFiles[0]}" extract; then
+      error "分层解压jar文件异常"
+    fi
+  fi
+
+  info "分层解压jar文件成功"
 
   # shellcheck disable=SC2164
   cd "${l_curDir}"
@@ -49,7 +63,7 @@ function _onBeforeCreatingDockerImage_ex() {
   for l_subDir in ${l_subDirs[@]};do
     #检查分层解压后的结果是否正确。
     if [[ ! -d "${gDockerBuildDir}/${l_subDir}" ]];then
-      error "命令(java -Djarmode=tools -jar ${l_jarFiles[0]} extract --destination ./extract --layers --launcher)执行异常：请检查${l_jarFiles[0]}包是否能正常运行"
+      error "命令(${l_command})执行后缺失${l_subDir}子目录,请检查${l_jarFiles[0]}包是否能正常运行"
     fi
     #判断子目录是否在l_dockerfile文件内容中出现，如果没有出现则删除之。
     # shellcheck disable=SC2002
